Crate chain_spec_builder

Substrate’s chain spec builder utility.

A chain-spec is short for chain-configuration. See the sc-chain-spec for more information.

Note that this binary is analogous to the build-spec subcommand, contained in typical substrate-based nodes. This particular binary is capable of building a more sophisticated chain specification that can be used with the substrate-node, ie. node-cli.

See ChainSpecBuilder for a list of available commands.

Enums

ChainSpecBuilder

A utility to easily create a testnet chain spec definition with a given set of authorities and endowed accounts and/or generate random accounts.

Functions

generate_authority_keys_and_store

Generate the authority keys and store them in the given keystore_path.

generate_chain_spec

Generate the chain spec using the given seeds and accounts.

print_seeds

Print the given seeds